refactor: doctor 유효성 검사 강화

+++ backend/KKUKKKUK/src/main/java/com/be/KKUKKKUK/domain/doctor/service/DoctorService.java
@@ -37,7 +37,7 @@ public class DoctorService {
     private final DoctorMapper doctorMapper;
 
     /**
-     * 특정 동물병원에 수의사를 등록합니다.
+     * 특정 동물병원에 새로운 수의사를 등록합니다.
      * @param hospital 조회 요청한 동물병원
      * @param request 등록할 수의사 정보
      * @return 등록 완료된 수의사 정보
@@ -47,17 +47,16 @@ public class DoctorService {
     }
 
     /**
-     * 특정 동물병원에 등록된 수의사 전체 목록을 조회합니다.
+     * 특정 수의사를 조회합니다.
      * 수의사가 속한 동물병원 계정만 의사 정보를 조회할 수 있습니다.
      * @param hospitalId 조회 요청한 동물병원 ID
+     * @param doctorId  조회할 수의사 ID
      * @return 조회된 수의사 정보
      */
     @Transactional(readOnly = true)
     public DoctorInfoResponse getDoctorInfoById(Integer hospitalId, Integer doctorId) {
         Doctor doctor = getDoctorById(doctorId);
-
         checkPermissionToDoctor(doctor, hospitalId);
-
         return doctorMapper.mapToDoctorInfo(doctor);
     }
 
@@ -73,16 +72,20 @@ public class DoctorService {
     }
 
     /**
-     * 등록된 수의사 정보를 삭제합니다.
+     * 등록된 특정 수의사 정보를 삭제합니다.
      * 수의사가 속한 동물병원 계정만 의사 정보를 삭제할 수 있습니다.
+     * 동물병원의 수의사가 두 명 이상일 때만 삭제가 가능합니다.
+     *
      * @param hospitalId 삭제 요청한 동물병원 ID
      * @param doctorId 삭제할 수의사 ID
      */
-    @Transactional
     public void deleteDoctorOnHospital(Integer hospitalId, Integer doctorId) {
         Doctor doctor = getDoctorById(doctorId);
-
         checkPermissionToDoctor(doctor, hospitalId);
+
+        if(getDoctorsByHospitalId(hospitalId).size() <= 1){
+            throw new ApiException(ErrorCode.DOCTOR_MINIMUM_LIMIT);
+        }
 
         doctorRepository.delete(doctor);
     }
@@ -96,7 +99,6 @@ public class DoctorService {
      * @param request 수정 요청
      * @return 수정 완료된 수의사 정보
      */
-    @Transactional
     public DoctorInfoResponse updateDoctorOnHospital(Integer hospitalId, Integer doctorId, DoctorUpdateRequest request) {
         Doctor doctor = getDoctorById(doctorId);
         checkPermissionToDoctor(doctor, hospitalId);
@@ -106,9 +108,7 @@ public class DoctorService {
     }
 
     /**
-     *
      * 수의사 ID 로 수의사 정보를 조회합니다.
-     *
      * @param doctorId 조회할 수의사 ID
      * @return Doctor entity
      * @throws ApiException 수의사를 찾을 수 없을 시 예외처리
@@ -122,6 +122,7 @@ public class DoctorService {
      * 동물병원이 특정 수의사에 대한 권한이 있는지 겁사하는 메소드
      * @param doctor 요청에 대한 수의사 entity
      * @param hospitalId 요청 시도한 동물병원 ID
+     * @throws ApiException 의사 소속의 병원이 아닌 경우 예외 발생
      */
     private void checkPermissionToDoctor(Doctor doctor, Integer hospitalId) {
         if (!doctor.getHospital().getId().equals(hospitalId)) {
